+++ alicompat/CHANGELOG.md
@@ -1,3 +1,10 @@
+## []
+
+- Added More Red support
+- Added Repurposed Structures support
+- Added Sophisticated Backpacks support
+- Added Sophisticated Storage support
+
 ## [1.0.0]
 
 - Initial release